BE completed -v1

+++ server/controllers/authorController.js
@@ -8,13 +8,14 @@ const ApiError = require('../error/ApiError');
 class AuthorController {
     async create(req, res, next) {
         try {
-            let {name, description, userId, countryId} = req.body;
+            const {id} = req.user;
+            let {name, description, countryId} = req.body;
             const {photo} = req.files;
             let fileName = uuid.v4() + ".jpg";
 
             photo.mv(path.resolve(__dirname, '..', 'static', fileName));
 
-            const author = await Author.create({name, description, userId, countryId, photo: fileName});            
+            const author = await Author.create({name, description, userId: id, countryId, photo: fileName});            
             return res.json(author);
 
         } catch(err) {
@@ -24,16 +25,17 @@ class AuthorController {
 
     async getAll(req, res, next) {
         try {
+            const {id} = req.user;
             let {countryId, limit, page} = req.query;  //req.query - в строку запроса. req.body - в теле запроса
             page = page || 1;
             limit = limit || 8;
             let offset = page * limit - limit;
-            let authors;
+            let authors;            
 
             if (countryId) {
-                authors = await Author.findAndCountAll({where:{countryId}, limit, offset});
+                authors = await Author.findAndCountAll({where:{userId: id, countryId}, limit, offset});
             } else {
-                authors = await Author.findAndCountAll({limit, offset});
+                authors = await Author.findAndCountAll({where:{userId: id}, limit, offset});
             }
 
             return res.json(authors);
@@ -46,7 +48,7 @@ class AuthorController {
     async getONe(req, res, next) {
         try {
             const {id} = req.params;
-            const author = await Author.findOne({where: {id}});
+            const author = await Author.findOne({where: {userId: req.user.id, id}});
             return res.json(author);
 
         } catch(err) {
@@ -57,7 +59,7 @@ class AuthorController {
     async delete(req, res, next) {
         try {
             const {id} = req.params;
-            await Author.destroy({where: {id}});
+            await Author.destroy({where: {userId: req.user.id, id}});
             return res.json('Author was deleted');
 
         } catch(err) {
@@ -68,8 +70,8 @@ class AuthorController {
     async update(req, res, next) {
         try {
             const {id} = req.params;
-            const {name, description, photo} = req.body;
-            await Author.update({name, description, photo}, {where: {id}});
+            const {name, description, photo, countryId} = req.body;
+            await Author.update({name, description, photo, countryId}, {where: {userId: req.user.id, id}});
             return res.json('Author was updated');
 
         } catch(err) {
