+++ server/controllers/bookController.js
@@ -8,13 +8,14 @@ const ApiError = require('../error/ApiError');
 class BookController {
     async create(req, res, next) {
         try {
-            let {name, link, rating, comment, userId, countryId, authorId} = req.body;
+            const {id} = req.user;
+            let {name, link, rating, comment, countryId, authorId} = req.body;
             const {cover} = req.files;
             let fileName = uuid.v4() + ".jpg";
 
             cover.mv(path.resolve(__dirname, '..', 'static', fileName));
 
-            const book = await Book.create({name, link, rating, comment, userId, countryId, authorId, cover: fileName});            
+            const book = await Book.create({name, link, rating, comment, userId: id, countryId, authorId, cover: fileName});            
             return res.json(book);
 
         } catch(err) {
@@ -24,6 +25,7 @@ class BookController {
 
     async getAll(req, res, next) {
         try {
+            const {id} = req.user;
             let {countryId, authorId, limit, page} = req.query;  //req.query - в строку запроса. req.body - в теле запроса
             page = page || 1;
             limit = limit || 8;
@@ -31,19 +33,19 @@ class BookController {
             let books;
 
             if (!countryId && !authorId) {
-                books = await Book.findAndCountAll({limit, offset});
+                books = await Book.findAndCountAll({where:{userId: id}, limit, offset});
             }
 
             if (countryId && !authorId) {
-                books = await Book.findAndCountAll({where:{countryId}, limit, offset});
+                books = await Book.findAndCountAll({where:{userId: id, countryId}, limit, offset});
             }
 
             if (!countryId && authorId) {
-                books = await Book.findAndCountAll({where:{authorId}, limit, offset});
+                books = await Book.findAndCountAll({where:{userId: id, authorId}, limit, offset});
             }
 
             if (countryId && authorId) {
-                books = await Book.findAndCountAll({where:{countryId, authorId}, limit, offset});
+                books = await Book.findAndCountAll({where:{userId: id, countryId, authorId}, limit, offset});
             }
 
             return res.json(books);
@@ -56,7 +58,7 @@ class BookController {
     async getONe(req, res, next) {
         try {
             const {id} = req.params;
-            const book = await Book.findOne({where: {id}});
+            const book = await Book.findOne({where: {userId: req.user.id, id}});
             return res.json(book);
 
         } catch(err) {
@@ -67,7 +69,7 @@ class BookController {
     async delete(req, res, next) {
         try {
             const {id} = req.params;
-            await Book.destroy({where: {id}});
+            await Book.destroy({where: {userId: req.user.id, id}});
             return res.json('Book was deleted');
 
         } catch(err) {
@@ -78,8 +80,8 @@ class BookController {
     async update(req, res, next) {
         try {
             const {id} = req.params;
-            const {name, link, rating, comment, cover} = req.body;
-            await Book.update({name, link, rating, comment, cover}, {where: {id}});
+            const {name, link, rating, comment, cover, countryId, authorId} = req.body;
+            await Book.update({name, link, rating, comment, cover, countryId, authorId}, {where: {userId: req.user.id, id}});
             return res.json('Book was updated');
 
         } catch(err) {
