+++ server/controllers/countryController.js
@@ -3,31 +3,33 @@ const ApiError = require('../error/ApiError');
 
 class CountryController {
     async create(req, res, next) {
+        const {id} = req.user;
         const {name} = req.body;
-        const candicate = await Country.findOne({where: {name}});
+        const candicate = await Country.findOne({where: {userId: id, name}});
         if (candicate) {
             return next(ApiError.badRequest('Такая страна уже существует!'));
         } 
 
-        const country = await Country.create({name});        
+        const country = await Country.create({userId: id, name});        
         return res.json(country);
     }
 
     async getAll(req, res) {
-        const countries = await Country.findAll();
+        const {id} = req.user;
+        const countries = await Country.findAll({where:{userId: id}});
         return res.json(countries);
     }
 
     async delete(req, res) {
         const {name} = req.params;
-        await Country.destroy({where: {name}});
+        await Country.destroy({where: {userId: req.user.id, name}});
         return res.json('Country was deleted');
     }
 
     async update(req, res) {
         const {id} = req.params;
         const {name} = req.body;
-        await Country.update({name}, {where: {id}});
+        await Country.update({name}, {where: {userId: req.user.id, id}});
         return res.json('Country was updated');
     }
 };
